Revert "update to vs2019"
This reverts commit 11a7ecae1dee4865e7e94c7f389f372c41d080de.

+++ BOSS/source/BuildOrderSearchGoal.cpp
@@ -140,7 +140,7 @@ bool BuildOrderSearchGoal::isAchievedBy(const GameState & state)
             }
         }
 
-        if (have < (int)getGoal(actionType))
+        if (have < getGoal(actionType))
         {
             return false;
         }
